Split the monolith into hooks
